Added privileges service
