+++ api/service/privileges.py
@@ -1,0 +1,28 @@
+from flask import request
+from api import db
+from api.model.database.users import Privilegio
+
+#TODO: separar POST e GET
+#TODO: remover verificação de método
+#TODO: remover verificação de json POST
+def Privileges():
+    if request.method == 'POST':
+        if request.is_json:
+            data = request.get_json()
+            new_privilege = Privilegio(user_type=data['user_type'])
+
+            db.session.add(new_privilege)
+            db.session.commit()
+
+            return {"message": f"Privilégio criado com sucesso"}
+        else:
+            return {"error": "A requisição não foi feita no formato esperado"}
+
+    elif request.method == 'GET':
+        privileges = Privilegio.query.all()
+        results = [
+            {
+                "user_type": privilege.user_type
+            } for privilege in privileges]
+
+        return {"count": len(results), "Privileges": results, "message": "success"}
